test: Migrate process_skill tests to packager module
- Move TestProcessSkill test class from agents test file to packager test file
- Update import statements and test function references
- Maintain existing test scenarios and assertions
- Improve test organization by keeping skill processing tests with related module
